Move initLora to RadioInterface.h
